nadia: xlrd xlwt: Import a subset of xlrd and xlwt sources
This patch creates the directory nadia -- the one that ought to contain
all the sources of Nadia -- and populates it with the subset of xlrd and
xlwt sources we need.

This patch is the first one in the branch that aims to split the unique
source file nadia.py into submodules and aims to document the behavior a
bit, in order to reduce complexity.  The branch is composed of the following
patches:

1. nadia: xlrd xlwt: Import a subset of xlrt and xlwt sources
2. nadia/django.py: Import some bits from the django project
3. nadia/__init__.py: Add initializer file for the Nadia module
4. nadia/excel.py: Extract from nadia.py code to read excel files
5. nadia/data.py: Extract from nadia.py code to read the data
6. nadia/yaxis.py: Extract from nadia.py code to read the Y axis
7. nadia/yaxis.py: "slugify" the machine readable name
8. nadia/finalize.py: Extract code to finalize data processing
9. nadia/jsonize.py: Add code to JSONize the dataset
10. nadia/main.py: Extract from nadia.py the "main" function

More in detail, this patch imports a subset of xlrd-0.7.1 and a subset
of <xlwt-0.7.2>.  More information on these two packages could be found
at <http://www.python-excel.org/>.