COPYING: Add 2011 to copyright statement
This patch adds 2011 to copyright statement because some code has been
written in 2011.

This is the first patch in a branch that aims to update documentation
and examples and tidy the root directory up.  The branch is composed of
the following patches:

1. COPYING: Add 2011 to copyright statement
2. Move documentation to doc/
3. Move testing files to test/
4. README: Reference doc/HOWTO rather than HOWTO

Follow-up branches will refactor the source code and split nadia.py
into smaller modules.