bin/nadia: Executable that locates nadia and invokes nadia.main()
If the program is not installed the content of the variable PREFIX is
ignored and we use "." as prefix.  This means that you must invoke nadia
as follows::

  ./bin/nadia ...

The installer should rewrite the content of bin/nadia and replace the
word "@PREFIX@" with the actual prefix where nadia is installed, for
example ``/usr/local``.