test: Update the conf files
This patch updates the configuration files after the recent changes
to ensure that they reference the right files and that the result is
the one we expect.

This is the first patch in a branch that aims to clean up stuff after
the recent huge amount of changes.  More in detail, the branch contains
the following patches:

1. test: Update the conf files
2. Remove nadia.py xlrd-0.7.1 xlwt-0.7.2
3. README: Note that now we also need json

A follow-up branch will attempt to clean-up the code a bit more (in
particular in nadia/jsonize.py) and, after that, I will try to implement
guessing, in order to avoid the need to compile at hand configuration
files.